Renaming folder and namespace domino.models
