+++ domino.models/domino.models.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
-    <RootNamespace>domino_models</RootNamespace>
+    <RootNamespace>domino.models</RootNamespace>
   </PropertyGroup>
 
 </Project>
